pyproject.toml
[project]
name = "opensec-env"
version = "0.1.0"
description = "OpenSec-Env OpenEnv-compatible cyber incident environment"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn>=0.24.0",
  "pydantic>=2.0.0",
  "requests>=2.25.0",
  "openai>=1.0.0",
  "openenv-core>=0.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0",
  "jsonschema>=4.0.0",
  "pyyaml>=6.0.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["server", "client", "sim", "oracle"]
exclude = ["data", "schemas", "notebooks", "docs", "scripts", "tests"]
